move relaxng go files to folder and update imports
