--- .travis.yml
@@ -1,12 +1,0 @@
-before_install:
-  - ./install-dependencies.sh
-
-install: true
-
-script:
-  - make travis
-
-language: go
-
-go:
-  - 1.7.1
