+++ genjssuite.sh
@@ -1,1 +1,1 @@
-go test -v -run=TestGenJSSuite 2> ./play/tests.js
+(cd relanxg && go test -v -run=TestGenJSSuite 2> ../play/tests.js)
