+++ go.mod
@@ -1,0 +1,16 @@
+module github.com/katydid/validator-go-relaxng
+
+go 1.24
+
+tool github.com/awalterschulze/checklicense
+
+require (
+	github.com/gopherjs/gopherjs v1.17.2
+	github.com/katydid/parser-go v0.7.2
+	github.com/katydid/parser-go-xml v0.1.1
+	github.com/katydid/validator-go v0.5.4
+)
+
+require (
+	github.com/awalterschulze/checklicense v1.0.1 // indirect
+)
