+++ relaxng/funcs.go
@@ -20,9 +20,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/katydid/katydid/relapse/ast"
-	c "github.com/katydid/katydid/relapse/combinator"
-	"github.com/katydid/katydid/relapse/funcs"
+	"github.com/katydid/validator-go/validator/ast"
+	c "github.com/katydid/validator-go/validator/combinator"
+	"github.com/katydid/validator-go/validator/funcs"
 )
 
 func stripTextPrefix(s string) (string, error) {
