+++ relaxng/relaxng.go
@@ -12,26 +12,27 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-//Translates RelaxNG to Katydid Relapse
+// Translates RelaxNG to Katydid Relapse
 package relaxng
 
 import (
 	"errors"
-	"github.com/katydid/katydid/parser/xml"
-	"github.com/katydid/katydid/relapse/ast"
-	"github.com/katydid/katydid/relapse/interp"
 	"reflect"
+
+	"github.com/katydid/parser-go-xml/xml"
+	"github.com/katydid/validator-go/validator/ast"
+	"github.com/katydid/validator-go/validator/interp"
 )
 
-//Translates a parsed RelaxNG Grammar into a Katydid Relapse Grammar.
+// Translates a parsed RelaxNG Grammar into a Katydid Relapse Grammar.
 func Translate(g *Grammar) (*ast.Grammar, error) {
 	return translate(g)
 }
 
-//The function removes the ns attributes with value TODO.
-//These ns="TODO" attributes can become present
-//after converting from RelaxNG to Simplified RelaxNG
-//using rng2srng.jar
+// The function removes the ns attributes with value TODO.
+// These ns="TODO" attributes can become present
+// after converting from RelaxNG to Simplified RelaxNG
+// using rng2srng.jar
 func RemoveTODOs(g *Grammar) {
 	removeTODOs(reflect.ValueOf(g).Elem())
 }
@@ -40,7 +41,7 @@ func NewXMLParser() xml.XMLParser {
 	return xml.NewXMLParser(xml.WithAttrPrefix("attr_"), xml.WithElemPrefix("elem_"), xml.WithTextPrefix("text_"))
 }
 
-//Validates input xml against a Katydid Relapse Grammar.
+// Validates input xml against a Katydid Relapse Grammar.
 func Validate(katydid *ast.Grammar, xmlContent []byte) error {
 	p := NewXMLParser()
 	if err := p.Init(xmlContent); err != nil {
